Update the code to split the modules
